requirements.txt
asgiref==3.3.4
Django==3.2.3
django-filter==2.4.0
django-render-partial==0.4
djangorestframework==3.12.4
Markdown==3.3.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
